added FutureWarning to empty Series without dtype and adjusted the tests and docs so that no unnecessary warnings are thrown

+++ pandas/tests/io/formats/test_format.py
@@ -2765,7 +2765,7 @@ class TestSeriesFormatting:
         assert res == exp
 
     def test_to_string_na_rep(self):
-        s = pd.Series(index=range(100))
+        s = pd.Series(index=range(100), dtype=np.float64)
         res = s.to_string(na_rep="foo", max_rows=2)
         exp = "0    foo\n      ..\n99   foo"
         assert res == exp
